API_gateway/requirements.txt

Flask
twilio
requests
google-auth
python-dotenv
